move: 폴더 이동 및 불필요한 파일 삭제
